Do not reset buffer on socket termination
The buffer in the consumer keeps track of which messages have been
processed by the subscriber. It only needs to be reset when there's a
change in the subscriber. If the socket gets closed due to e.g. some
network problem, the subscriber may still keep processing and ack-ing
the messages already fetched.