Always clean up after the old subscriber before accepting a new one
This essentially moves the `reset_buffer` call before the
`update_options`, and has the same effect as handling the `DOWN`
message from the old subscriber before the subscribe request from the
new one (which should be the typical case).

The reason is that the `begin_offset` specified in the subscribe
request (if present) shall take precedence over the automatic update
of this field in `reset_buffer`.