Support limiting the initial prefetch window size in `brod_consumer`
This feature is most useful when the `brod_consumer` process crashes,
but the subscriber process survives and attempts to subscribe again
once the `brod_consumer` is restarted. In this case the subscriber may
still have some unacked messages received from the crashed
`brod_conumser`. It needs to restart from the last offset + 1 it has
seen, so it won't receive duplicated messages, but also needs to
reduce the initial prefetch window, otherwise a frequently crashing
`brod_consumer` may flood it with messages.