Fix wrong handeling of empty fetch response.
Empty fetch response happens only in incremental fetch sessions.
The current parsing code is a copy-paste from kafka_protocol test_lib code
which is fine.
But the comment is wrong: It's fetch session not transactional session
which is a entirely different thing.
And the handling is even more wrong: In case empty fetch response is
received, it should not advance begin_offset, but keep retrying the one
used in the last fetch request, because essentially an empty
response implies no new offsets and no partition metadata change either.