Receive pending acks after assignments_revoked is invoked
We are building a multi-process data processing pipeline
with Brod and we are using the assignments_revoked callback
to flush all messages being processed. However, if any message
is acknowledged during the assignments_revoked step, those
acks won't be committed, they will just sit in the group coordinator
inbox and eventually be processed too late.

This commit makes it so we retrieve any pending ack from the
inbox before we commit them.