Fix the exponential growth of the producer buffer
Referencing `State#state.partition` from within the `BufCb` fun adds
the entire state (including the producer buffer) to the free variables
of the fun. Then the fun is added to the producer buffer and in turn
to the new version of the state.

This means that when multiple produce requests are buffered, the
"flattened size" of the producer buffer will grow exponentially. The
observable memory growth of the producer process will be linear,
because the old and the new state reference a lot of common terms on
the heap. However, when the producer buffer is copied (e.g. when
sending it in a message) all those duplicate references lead to
creating duplicate copies of the referenced terms, hence the memory
need of the message will grow exponentially.

This problem can be easily avoided by making sure only the necessary
`partition` field of the state becomes a free variable of the fun.