Remove validate_only option from create_topics spec
The validate_only option is ignored as it was introduced
in API version 1 and brod still uses version 0 for the
create_topics request.

It is thus confusing to document it as a valid option.

This commit may be reverted when brod starts to use the
V1 API (or newer) for create_topics.

See: brod_kafka_apis.erl:155, kpro_req_lib.erl:329