Use h2-h4 in README
The top level headings in Markdown (#) correspond
to HTML h1 tags and should be used only for a single
heading.

ExDoc shows h2 headings in the left navigation bar,
so when using h1 for 2nd level, there are many
irrelevant navigation items and the relevant ones
are missing. Also HTML ids are not generated for h1
headings in ExDoc so one cannot link to a particular
heading.

I left the Notice as h1 though, as I assume you want
to keep it as visible as possible.