feature: support for transactions
Now, `brod_producer` sends the transaction id if needed. To do that, the changes here add a parameter on init and bypass the producer register. Since producers are topic-partition-bounded but a transaction can spread over many topic-partitions. This commit adds a new module called `brod_transaction` to orchestrate them and offer an API to interact with transactions. Finally, the module `brod` adds the API to the users.

This commit includes a set of tests for the new functionality.